+++ cstar/tests/unit_tests/marbl/test_marbl_external_codebase.py
@@ -1,5 +1,6 @@
 import os
 import pathlib
+from typing import Any
 from unittest import mock
 
 import dotenv
@@ -10,7 +11,7 @@ from cstar.system.manager import cstar_sysmgr
 
 
 @pytest.fixture
-def marbl_codebase():
+def codebase():
     """Fixture providing a configured instance of `MARBLExternalCodeBase` for
     testing."""
     source_repo = "https://github.com/marbl-ecosys/MARBL.git"
@@ -20,22 +21,19 @@ def marbl_codebase():
     )
 
 
-def test_default_source_repo(marbl_codebase):
+def test_default_source_repo(codebase):
     """Test if the default source repo is set correctly."""
-    assert (
-        marbl_codebase.default_source_repo
-        == "https://github.com/marbl-ecosys/MARBL.git"
-    )
+    assert codebase.default_source_repo == "https://github.com/marbl-ecosys/MARBL.git"
 
 
-def test_default_checkout_target(marbl_codebase):
+def test_default_checkout_target(codebase):
     """Test if the default checkout target is set correctly."""
-    assert marbl_codebase.default_checkout_target == "marbl0.45.0"
+    assert codebase.default_checkout_target == "marbl0.45.0"
 
 
-def test_expected_env_var(marbl_codebase):
+def test_expected_env_var(codebase):
     """Test if the expected environment variable is set correctly."""
-    assert marbl_codebase.expected_env_var == "MARBL_ROOT"
+    assert codebase.expected_env_var == "MARBL_ROOT"
 
 
 def test_defaults_are_set():
@@ -97,23 +95,24 @@ class TestMARBLExternalCodeBaseGet:
         self,
         dotenv_path: pathlib.Path,
         marbl_path: pathlib.Path,
-        marbl_codebase: MARBLExternalCodeBase,
+        codebase: MARBLExternalCodeBase,
     ):
         """Test that the get method succeeds when subprocess calls succeed."""
         # Setup:
         ## Mock success of calls to subprocess.run:
         self.mock_subprocess_run.return_value.returncode = 0
 
-        key = marbl_codebase.expected_env_var
+        key = codebase.expected_env_var
         value = str(marbl_path)
 
         with mock.patch(
-            "cstar.system.environment.CSTAR_USER_ENV_PATH",
-            dotenv_path,
+            "cstar.system.environment.CStarEnvironment.user_env_path",
+            new_callable=mock.PropertyMock,
+            return_value=dotenv_path,
         ):
             # Test
             ## Call the get method
-            marbl_codebase.get(target=marbl_path)
+            codebase.get(target=marbl_path)
 
             # Assertions:
             ## Check environment variables
@@ -121,9 +120,9 @@ class TestMARBLExternalCodeBaseGet:
 
             ## Check that _clone_and_checkout was (mock) called correctly
             self.mock_clone_and_checkout.assert_called_once_with(
-                source_repo=marbl_codebase.source_repo,
+                source_repo=codebase.source_repo,
                 local_path=marbl_path,
-                checkout_target=marbl_codebase.checkout_target,
+                checkout_target=codebase.checkout_target,
             )
 
             ## Check that environment was updated correctly
@@ -138,7 +137,7 @@ class TestMARBLExternalCodeBaseGet:
                 shell=True,
             )
 
-    def test_make_failure(self, marbl_codebase, tmp_path):
+    def test_make_failure(self, codebase, tmp_path):
         """Test that the get method raises an error when 'make' fails."""
 
         ## There are two subprocess calls, we'd like one fail, one pass:
@@ -158,8 +157,61 @@ class TestMARBLExternalCodeBaseGet:
                 ),
             ),
             mock.patch(
-                "cstar.system.environment.CSTAR_USER_ENV_PATH",
-                dotenv_path,
+                "cstar.system.environment.CStarEnvironment.user_env_path",
+                new_callable=mock.PropertyMock,
+                return_value=dotenv_path,
             ),
         ):
-            marbl_codebase.get(target=tmp_path)
+            codebase.get(target=tmp_path)
+
+    def test_local_config_status_with_prebuilt_and_root(
+        self,
+        codebase: MARBLExternalCodeBase,
+        custom_user_env: Any,
+    ):
+        """Test that the `local_config_status` method returns an appropriate return code
+        when a prebuilt MARBL codebase is configured but the expected environment
+        variable is not set."""
+
+        mock_path = "/any/path/it/is/not/verified"
+        custom_user_env(
+            {
+                codebase.prebuilt_env_var: "1",
+                codebase.expected_env_var: mock_path,
+            }
+        )
+        assert os.environ[codebase.prebuilt_env_var] == "1"
+        assert os.environ[codebase.expected_env_var] == mock_path
+
+        # Test
+        status_code = codebase.local_config_status
+
+        ## Assertions:
+        # Confirm that the status is 0 when prebuilt & expected var are set
+        assert status_code == 0
+
+    def test_local_config_status_with_prebuilt_and_no_root(
+        self,
+        codebase: MARBLExternalCodeBase,
+        custom_user_env: Any,
+    ):
+        """Test that the `local_config_status` method returns an appropriate return code
+        when a prebuilt MARBL codebase is configured but the expected environment
+        variable is not set."""
+
+        # Do NOT set `xxx_codebase.expected_env_var`, only prebuilt_env_var
+        custom_user_env({codebase.prebuilt_env_var: "1"})
+        assert os.environ[codebase.prebuilt_env_var] == "1"
+        assert codebase.expected_env_var not in os.environ
+
+        # Test
+        status_code = codebase.local_config_status
+
+        ## Assertions:
+        # Confirm that the status is 4 due to missing XXX_ROOT
+        assert status_code == 4
+
+    def test_prebuilt_env_var(self):
+        """Verify that the value of the prebuilt_env_var is set correctly."""
+        marbl_codebase = MARBLExternalCodeBase()
+        assert marbl_codebase.prebuilt_env_var == "CSTAR_MARBL_PREBUILT"